requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
requests==2.21.0
Pillow>=5.3.0,<5.4.0
rest-social-auth==2.1.0
django-cors-headers==2.5.2
coreapi==2.3.3
graphviz==0.10.1
django-extensions==2.1.6
pydot==1.4.1
pydotplus==2.0.2
#pygraphviz==1.5

gunicorn==19.9.0

flake8>=3.6.0,<3.7.0